Move config to its own file
